Resolvendo possiveis conflitos e expondo versao final do app
